main/requirements.txt
Django==3.2.6
django-environ==0.4.5
djangorestframework==3.12.4
requests==2.25.1
gunicorn==20.1.0